.serverless/poolingApplicationAPI/requirements.txt
attrs==24.2.0
boto3==1.34.155
botocore==1.34.155
jmespath==1.0.1
jsonschema==4.17.3
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
